noi anda el dto queee

+++ BackEnd/src/main/java/com/portfolio/demo/Interface/IPersonaService.java
@@ -12,10 +12,9 @@ public interface IPersonaService {
     public void savePersona(Persona persona);
     
     //Eliminar un usuario pero lo buscamos por id
-    public void deletePersona(Long id);
+    public void deletePersona(int id);
    
     //Buscar una persona por id
-    public Persona findPersona(Long id);
+    public Persona findPersona(int id);
     
 }
-
